Extract LOCProvider to its own file
